fixed ui/string warnings

+++ src/ui/string.h
@@ -1,8 +1,8 @@
 
 extern u16 str_to_u16(const char *string);
-extern char *u16_to_str(char *out, int outlen, u16 num);
-extern char *string_zero_pad(char *out, int outlen, const char *str, u16 pad_size);
-extern char *u16_to_hex(char *out, int outlen, u16 num);
+extern char *u16_to_str(char *out, size_t outlen, u16 num);
+extern char *string_zero_pad(char *out, size_t outlen, const char *str, u16 pad_size);
+extern char *u16_to_hex(char *out, size_t outlen, u16 num);
 extern char *string_reverse(char *str);
 extern char *string_lower(char *str);
